+++ tests/web_api/loader/test_parameters.py
@@ -1,10 +1,8 @@
 # -*- coding: utf-8 -*-
 
-import pytest
+from openfisca_core.parameters import Scale
 
-from openfisca_core.parameters import ParameterNode, Scale
-
-from openfisca_web_api.loader.parameters import build_api_scale, walk_node
+from openfisca_web_api.loader.parameters import build_api_scale, build_api_parameter
 
 
 def test_build_rate_scale():
@@ -21,30 +19,17 @@ def test_build_amount_scale():
     assert build_api_scale(rate, 'amount') == {'2014-01-01': {1: 0}}
 
 
-@pytest.fixture
-def walk_node_args():
-    parameters = []
-    path_fragments = []
-    country_package_metadata = {'location': 'foo', 'version': '1', 'repository_url': 'foo'}
-
-    return parameters, path_fragments, country_package_metadata
-
-
-def test_walk_node_rate_scale(walk_node_args):
-    '''Serializes a 'rate' parameter node'''
-    root_node = ParameterNode(data = {})
+def test_full_rate_scale():
+    '''Serializes a 'rate' scale parameter'''
     data = {'brackets': [{'rate': {'2014-01-01': {'value': 0.5}}, 'threshold': {'2014-01-01': {'value': 1}}}]}
-    scale = Scale('this rate', data, None)
-    root_node.children['rate'] = scale
-    parameters = walk_node(root_node, *walk_node_args)
-    assert parameters == [{'description': None, 'id': 'rate', 'metadata': {}, 'brackets': {'2014-01-01': {1: 0.5}}}]
+    scale = Scale('rate', data, None)
+    api_scale = build_api_parameter(scale, {})
+    assert api_scale == {'description': None, 'id': 'rate', 'metadata': {}, 'brackets': {'2014-01-01': {1: 0.5}}}
 
 
-def test_walk_node_amount_scale(walk_node_args):
-    '''Serializes an 'amount' parameter node'''
-    root_node = ParameterNode(data = {})
+def test_walk_node_amount_scale():
+    '''Serializes an 'amount' scale parameter '''
     data = {'brackets': [{'amount': {'2014-01-01': {'value': 0}}, 'threshold': {'2014-01-01': {'value': 1}}}]}
-    scale = Scale('that amount', data, None)
-    root_node.children['amount'] = scale
-    parameters = walk_node(root_node, *walk_node_args)
-    assert parameters == [{'description': None, 'id': 'amount', 'metadata': {}, 'brackets': {'2014-01-01': {1: 0}}}]
+    scale = Scale('amount', data, None)
+    api_scale = build_api_parameter(scale, {})
+    assert api_scale == {'description': None, 'id': 'amount', 'metadata': {}, 'brackets': {'2014-01-01': {1: 0}}}