+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 24.9.0 [#767](https://github.com/openfisca/openfisca-core/pull/767)
+
+- Introduce `ParameterNode.get_descendants`
+  - This method returns a generator containing all the parameters and nodes recursively contained in the `ParameterNode`
+
 ### 24.8.2 [#773](https://github.com/openfisca/openfisca-core/pull/773)
 
 - Make sure to cap all dependency versions, in order to avoid unwanted functional and integration breaks caused by external code updates